:recycle:refactor: seperate types from book detail API 책 상세 페이지 API 타입 분리
